backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery[redis]
django-celery-beat
python-dotenv
pytest
pytest-django
requests
bcrypt
selenium
gunicorn
django-cors-headers